+++ geopackage-sdk/src/main/java/mil/nga/geopackage/tiles/overlay/FeatureOverlay.java
@@ -2,7 +2,11 @@ package mil.nga.geopackage.tiles.overlay;
 
 import com.google.android.gms.maps.model.Tile;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import mil.nga.geopackage.tiles.features.FeatureTiles;
+import mil.nga.geopackage.tiles.user.TileDao;
 
 /**
  * Feature overlay which draws tiles from a feature table
@@ -15,6 +19,11 @@ public class FeatureOverlay extends BoundedOverlay {
      * Feature tiles
      */
     private final FeatureTiles featureTiles;
+
+    /**
+     * Linked GeoPackage overlays
+     */
+    private List<GeoPackageOverlay> linkedOverlays = new ArrayList<>();
 
     /**
      * Constructor
@@ -39,6 +48,24 @@ public class FeatureOverlay extends BoundedOverlay {
      * {@inheritDoc}
      */
     @Override
+    public boolean hasTileToRetrieve(int x, int y, int zoom) {
+
+        // Determine if the tile should be drawn
+        boolean drawTile = true;
+        for (GeoPackageOverlay geoPackageOverlay : linkedOverlays) {
+            if (geoPackageOverlay.hasTile(x, y, zoom)) {
+                drawTile = false;
+                break;
+            }
+        }
+
+        return drawTile;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public Tile retrieveTile(int x, int y, int zoom) {
 
         // Draw the tile
@@ -51,6 +78,41 @@ public class FeatureOverlay extends BoundedOverlay {
         }
 
         return tile;
+    }
+
+    /**
+     * Ignore drawing tiles if they exist in the tile tables represented by the tile daos
+     *
+     * @param tileDaos tile data access objects
+     * @since 1.2.6
+     */
+    public void ignoreTileDaos(List<TileDao> tileDaos) {
+
+        for (TileDao tileDao : tileDaos) {
+            ignoreTileDao(tileDao);
+        }
+
+    }
+
+    /**
+     * Ignore drawing tiles if they exist in the tile table represented by the tile dao
+     *
+     * @param tileDao tile data access object
+     * @since 1.2.6
+     */
+    public void ignoreTileDao(TileDao tileDao) {
+
+        GeoPackageOverlay tileOverlay = new GeoPackageOverlay(tileDao);
+        linkedOverlays.add(tileOverlay);
+    }
+
+    /***
+     * Clear all ignored tile tables
+     *
+     * @since 1.2.6
+     */
+    public void clearIgnored() {
+        linkedOverlays.clear();
     }
 
 }