+++ geopackage-sdk/src/main/java/mil/nga/geopackage/tiles/overlay/GoogleAPIGeoPackageOverlay.java
@@ -33,6 +33,14 @@ public class GoogleAPIGeoPackageOverlay extends BoundedOverlay {
      * {@inheritDoc}
      */
     @Override
+    public boolean hasTileToRetrieve(int x, int y, int zoom) {
+        return retriever.hasTile(x, y, zoom);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public Tile retrieveTile(int x, int y, int zoom) {
 
         GeoPackageTile geoPackageTile = retriever.getTile(x, y, zoom);