+++ CHANGELOG.md
@@ -6,7 +6,11 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.2.6 (TBD)
 
-* TBD
+* geopackage-core version updated to 1.1.6
+* Feature Tile Table Linker implementation with methods for retrieving data access objects
+* Bounded Overlay check if a specified tile exists
+* Feature Overlay ignore drawing tiles that already exist in a linked tile table
+* Feature Overlay Query improved determination if a tile exists before querying features
 
 ## [1.2.5](https://github.com/ngageoint/geopackage-android/releases/tag/1.2.5) (02-02-2016)